Ajout des vérifications de droit pour le CRUD événement, validé via Postman

+++ src/main/java/com/fil_rouge_frontoffice/controller/EvenementRestController.java
@@ -3,8 +3,10 @@ package com.fil_rouge_frontoffice.controller;
 import com.fil_rouge_frontoffice.controller.dto.EvenementDto;
 
 import com.fil_rouge_frontoffice.entity.Evenement;
+import com.fil_rouge_frontoffice.entity.Utilisateur;
 import com.fil_rouge_frontoffice.service.EvenementService;
 
+import com.fil_rouge_frontoffice.service.UtilisateurService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,10 +22,15 @@ import java.util.Optional;
 public class EvenementRestController {
     @Autowired
     private EvenementService evenementService;
+    @Autowired
+    private UtilisateurService utilisateurService;
 
     @PostMapping("/add")
     public ResponseEntity<EvenementDto> addEvenement(@RequestBody EvenementDto dto){
-        //Utilisateur utilisateur = utilisateurService.getConnectedUtilisateur();
+        Utilisateur utilisateurConnecte = utilisateurService.getConnectedUtilisateur();
+        if(!utilisateurConnecte.getMail().equals(dto.getMailUtilisateur())){
+            if(!utilisateurService.isEcritureAutorisee(dto.getMailUtilisateur(), utilisateurConnecte.getMail())) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         EvenementDto evenementDto = evenementService.addEvenement(dto);
         return ResponseEntity.status(HttpStatus.CREATED).body(evenementDto);
     }
@@ -31,6 +38,11 @@ public class EvenementRestController {
     public ResponseEntity<HttpStatus> deleteEvenement(@PathVariable("id") Long idEvenement){
         Optional<Evenement> optEvenement = evenementService.findById(idEvenement);
         if(optEvenement.isPresent()){
+            Utilisateur utilisateurConnecte = utilisateurService.getConnectedUtilisateur();
+            Utilisateur proprietaire = optEvenement.get().getUtilisateur();
+            if(!utilisateurConnecte.getMail().equals(proprietaire.getMail())){
+                if(!utilisateurService.isSuppressionAutorisee(proprietaire.getMail(), utilisateurConnecte.getMail())) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
             evenementService.delete(idEvenement);
             return ResponseEntity.status(HttpStatus.OK).build();
         }
@@ -40,8 +52,16 @@ public class EvenementRestController {
     @PutMapping("/update")
     public ResponseEntity<HttpStatus> updateEvenement(@RequestBody EvenementDto evenement) {
 
-        boolean exists = evenementService.update(evenement);
-        if(exists) return ResponseEntity.status(HttpStatus.OK).build();
+        boolean exists = evenementService.existsById(evenement.getIdEvenement());
+        if(exists) {
+            Utilisateur utilisateurConnecte = utilisateurService.getConnectedUtilisateur();
+            Utilisateur proprietaire = evenementService.findById(evenement.getIdEvenement()).get().getUtilisateur();
+            if(!utilisateurConnecte.getMail().equals(proprietaire.getMail())){
+                if(!utilisateurService.isModificationAutorisee(proprietaire.getMail(), utilisateurConnecte.getMail())) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+            evenementService.update(evenement);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
     }
 
@@ -49,6 +69,11 @@ public class EvenementRestController {
     public ResponseEntity<EvenementDto> fetchEvenementById(@PathVariable("id") Long id){
         Optional<Evenement> evenementOpt = evenementService.findById(id);
         if(evenementOpt.isPresent()){
+            Utilisateur utilisateurConnecte = utilisateurService.getConnectedUtilisateur();
+            Utilisateur proprietaire = evenementOpt.get().getUtilisateur();
+            if(!utilisateurConnecte.getMail().equals(proprietaire.getMail())){
+                if(!utilisateurService.isLectureAutorisee(proprietaire.getMail(), utilisateurConnecte.getMail())) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
             EvenementDto dto = EvenementDto.from(evenementOpt.get());
             return ResponseEntity.status(HttpStatus.OK).body(dto);
         }
@@ -58,6 +83,10 @@ public class EvenementRestController {
 
     @GetMapping("/proprietaire/{mail}")
     public ResponseEntity<List<EvenementDto>> fetchEvenementsByMail(@PathVariable("mail") String mail) {
+        String mailUtilisateurConnecte = utilisateurService.getConnectedUtilisateur().getMail();
+        if(!mailUtilisateurConnecte.equals(mail)){
+            if(!utilisateurService.isLectureAutorisee(mail, mailUtilisateurConnecte)) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<EvenementDto> listeEvenements = evenementService.findUpcomingEvenementsUtilisateur(mail);
         return ResponseEntity.status(HttpStatus.OK).body(listeEvenements);
     }
